rename new project & update NuGet reference format
