+++ SkillPrestige.Magic/Framework/UpgradePointSpecialHandling.cs
@@ -7,26 +7,26 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SkillPrestige.MagicSkill.Framework
+namespace SkillPrestige.Magic.Framework
 {
     /// <summary>Special handling for adding to the mana cap from a Magic profession..</summary>
-    internal class ManaCapSpecialHandling : IProfessionSpecialHandling
+    internal class UpgradePointSpecialHandling : IProfessionSpecialHandling
     {
         public readonly int Amount;
 
-        public ManaCapSpecialHandling(int amount)
+        public UpgradePointSpecialHandling(int amount)
         {
             this.Amount = amount;
         }
 
         public void ApplyEffect()
         {
-            Game1.player.setMaxMana(Game1.player.getMaxMana() + Amount);
+            Game1.player.useSpellPoints(-Amount, true);
         }
 
         public void RemoveEffect()
         {
-            Game1.player.setMaxMana(Game1.player.getMaxMana() - Amount);
+            Game1.player.useSpellPoints(Amount, true);
         }
     }
 }
