+++ SkillPrestige.Magic/Properties/AssemblyInfo.cs
