--- SkillPrestige.MagicSkill/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Pathoschild.Stardew.ModBuildConfig" version="2.2.0" targetFramework="net452" />
-</packages>
