+++ SkillPrestige.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27703.2042
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28803.202
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkillPrestige", "SkillPrestige\SkillPrestige.csproj", "{D444AC8A-8D4C-4256-94BB-B075B717FEFF}"
 EndProject
@@ -13,7 +13,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkillPrestige.LuckSkill", "
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkillPrestige.CookingSkill", "SkillPrestige.CookingSkill\SkillPrestige.CookingSkill.csproj", "{A1415907-3D4B-4D55-AA4A-063696DF3019}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkillPrestige.MagicSkill", "SkillPrestige.MagicSkill\SkillPrestige.MagicSkill.csproj", "{B353BCD7-BB57-421D-BFF5-EFBA59B62FA8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkillPrestige.Magic", "SkillPrestige.Magic\SkillPrestige.Magic.csproj", "{B353BCD7-BB57-421D-BFF5-EFBA59B62FA8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
